osd: add option for "unscaled" OSD

+++ sub/ass_mp.c
@@ -39,7 +39,7 @@
 
 // res_y should be track->PlayResY
 // It determines scaling of font sizes and more.
-void mp_ass_set_style(ASS_Style *style, int res_y,
+void mp_ass_set_style(ASS_Style *style, double res_y,
                       const struct osd_style_opts *opts)
 {
     if (opts->font) {
